continued CSRF protection impl

+++ controllers/projects/mdelete.controller.php
@@ -13,21 +13,22 @@
 
         header("location: /");
     }
+    if ($_SERVER["REQUEST_METHOD"] == "POST" && hash_equals($_SESSION["csrf_token"], $_POST['csrf_token'])) {
+        $db = new Db('localhost', 'root1', 'root', 'projectsms');
 
-    $db = new Db('localhost', 'root1', 'root', 'projectsms');
+        $datas = $_POST['chkbox'];
 
-    $datas = $_POST['chkbox'];
-
-    foreach ($datas as $data) {
-        $delete = $db->delete('projects', 'id=?', array($data));
-        $delete = $db->delete('projects_users', 'project_id=?', array($data));
-        $delete = $db->delete('tasks', 'project=?', array($data));
-    }
-    if ($delete) {
-        echo "<div class='rounded-5 centerdiv h5 text-center shadow w-50 rounded-pill'> <br>
+        foreach ($datas as $data) {
+            $delete = $db->delete('projects', 'id=?', array($data));
+            $delete = $db->delete('projects_users', 'project_id=?', array($data));
+            $delete = $db->delete('tasks', 'project=?', array($data));
+        }
+        if ($delete) {
+            echo "<div class='rounded-5 centerdiv h5 text-center shadow w-50 rounded-pill'> <br>
          Projects and its realted tasks/users were Deleted Successfully <br>
          <img src='/views/imgs/icons/Gifdelete.gif' width='50' height='50'> </div>";
-        header("Refresh: 2; URL=/");
-    } else {
-        echo "Error";
+            header("Refresh: 2; URL=/");
+        } else {
+            echo "Error";
+        }
     }
